requirements.txt
torch
torchvision
torchaudio
gradio==4.20.0
flask
numpy
pandas
scikit-learn
matplotlib
seaborn
pillow
snntorch
spikingjelly
langchain
langchain-community
openai
transformers
sentence-transformers
datasets
omegaconf
hydra-core
dependency-injector
tqdm
jupyter
ipywidgets
requests
beautifulsoup4
typer==0.16.0
PyYAML
h5py
GPyOpt
GPy
matplotlib
cupy==13.6.0
networkx
pandas-stubs
